Updated deep_sort and fixed import paths from update
