+++ .idea/vcs.xml
@@ -2,7 +2,5 @@
 <project version="4">
   <component name="VcsDirectoryMappings">
     <mapping directory="$PROJECT_DIR$" vcs="Git" />
-    <mapping directory="$PROJECT_DIR$/deep_sort" vcs="Git" />
-    <mapping directory="$PROJECT_DIR$/umt/deep_sort" vcs="Git" />
   </component>
 </project>
