+++ .idea/workspace.xml
@@ -26,8 +26,31 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="2f6d1c78-ac22-44a2-9611-affb608006f6" name="Default Changelist" comment="">
-      <change afterPath="$PROJECT_DIR$/umt/deep_sort/__init__.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/umt/umt_main.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/umt_main.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/umt/deep_sort/LICENSE" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/umt/deep_sort/application_util/__init__.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/umt/deep_sort/deep_sort/__init__.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/umt/deep_sort/deep_sort_app.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/umt/deep_sort/deep_sort_tools/__init__.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/umt/deep_sort/deep_sort_tools/freeze_model.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/umt/deep_sort/evaluate_motchallenge.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/umt/deep_sort/generate_videos.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/umt/deep_sort/requirements.txt" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/umt/deep_sort/setup.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/umt/deep_sort/show_results.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/deep_sort/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/detection.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/deep_sort/deep_sort/detection.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/generate_detections.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/deep_sort/deep_sort_tools/generate_detections.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/image_viewer.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/deep_sort/application_util/image_viewer.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/iou_matching.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/deep_sort/deep_sort/iou_matching.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/kalman_filter.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/deep_sort/deep_sort/kalman_filter.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/linear_assignment.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/deep_sort/deep_sort/linear_assignment.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/mars-small128.pb" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/nn_matching.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/deep_sort/deep_sort/nn_matching.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/preprocessing.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/deep_sort/application_util/preprocessing.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/track.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/deep_sort/deep_sort/track.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/tracker.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/deep_sort/deep_sort/tracker.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/umt/deep_sort/visualization.py" beforeDir="false" afterPath="$PROJECT_DIR$/umt/deep_sort/application_util/visualization.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -58,6 +81,7 @@
   <component name="PropertiesComponent">
     <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
     <property name="RunOnceActivity.ShowReadmeOnStart" value="true" />
+    <property name="settings.editor.selected.configurable" value="project.propVCSSupport.Mappings" />
   </component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
@@ -82,7 +106,14 @@
       <option name="project" value="LOCAL" />
       <updated>1615379241714</updated>
     </task>
-    <option name="localTasksCounter" value="2" />
+    <task id="LOCAL-00002" summary="Updated deep_sort and fixed import paths from update">
+      <created>1615386326423</created>
+      <option name="number" value="00002" />
+      <option name="presentableId" value="LOCAL-00002" />
+      <option name="project" value="LOCAL" />
+      <updated>1615386326423</updated>
+    </task>
+    <option name="localTasksCounter" value="3" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -109,13 +140,22 @@
     </option>
   </component>
   <component name="VcsManagerConfiguration">
+    <ignored-roots>
+      <path value="$PROJECT_DIR$/deep_sort" />
+      <path value="$PROJECT_DIR$/umt/deep_sort" />
+    </ignored-roots>
     <MESSAGE value="Possible fix for the get_class error" />
-    <option name="LAST_COMMIT_MESSAGE" value="Possible fix for the get_class error" />
+    <MESSAGE value="Updated deep_sort and fixed import paths from update" />
+    <option name="LAST_COMMIT_MESSAGE" value="Updated deep_sort and fixed import paths from update" />
   </component>
   <component name="WindowStateProjectService">
-    <state x="873" y="433" key="Vcs.Push.Dialog.v2" timestamp="1615379305996">
+    <state x="782" y="345" key="SettingsEditor" timestamp="1615387142235">
       <screen x="0" y="0" width="2560" height="1400" />
     </state>
-    <state x="873" y="433" key="Vcs.Push.Dialog.v2/0.0.2560.1400@0.0.2560.1400" timestamp="1615379305996" />
+    <state x="782" y="345" key="SettingsEditor/0.0.2560.1400@0.0.2560.1400" timestamp="1615387142235" />
+    <state x="873" y="433" key="Vcs.Push.Dialog.v2" timestamp="1615386332983">
+      <screen x="0" y="0" width="2560" height="1400" />
+    </state>
+    <state x="873" y="433" key="Vcs.Push.Dialog.v2/0.0.2560.1400@0.0.2560.1400" timestamp="1615386332983" />
   </component>
 </project>
