--- umt/deep_sort/__init__.py
@@ -1,1 +1,0 @@
-# vim: expandtab:ts=4:sw=4
