+++ umt/deep_sort/requirements.txt
@@ -1,0 +1,3 @@
+tensorflow==2.4.0
+numpy==1.19.5
+opencv-python==4.5.1.48
